fix: refactor dropdown filter, sticky header
