+++ packages/modules/data-widgets/src/themesource/datawidgets/web/_datagrid-filters.scss
@@ -229,8 +229,6 @@ $item-min-height: 32px;
     }
 
     .dropdown-list {
-        position: absolute;
-        min-width: 100%;
         left: 0;
         margin: 0 $spacing-small;
         padding: 0;
