+++ packages/pluggableWidgets/datagrid-dropdown-filter-web/src/utils/helper.ts
@@ -1,0 +1,3 @@
+export function preventReactErrorsAboutReadOnly(): void {
+    return undefined;
+}
