+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   We removed redundant code to improve widget load time in the browser.
 -   We fixed an issue where column selector and checkbox icon not aligned with filter control.
+-   We fixed issues with filter position overlapping below the next container.
+-   We changed DOM structure to allow sticky header on virtual scrolling pagination mode.
 
 ## [2.8.2] - 2023-08-25
