+++ packages/shared/widget-plugin-grid/src/components/StickyHeaderTable.tsx
@@ -1,11 +1,16 @@
-import { createElement, ReactElement, useCallback, useEffect, useRef, PropsWithChildren, Fragment } from "react";
+import classNames from "classnames";
+import { PropsWithChildren, ReactElement, createElement, useCallback, useEffect, useRef } from "react";
+import { InfiniteBodyProps, useInfiniteControl } from "./InfiniteBody";
 
 // this wrapper component works by adding "sticky-sentinel" div (and hides it using margin-bottom -1px)
 // in order to detect table going out of viewport
 // this allows position: sticky to only applies when container is "stuck"
 // this way we allow any absolute position to still be on top when sticky not "stuck"
-export function StickyHeaderTable(props: PropsWithChildren): ReactElement {
+export function StickyHeaderTable(props: PropsWithChildren<InfiniteBodyProps>): ReactElement {
+    const { className, isInfinite, children } = props;
     const headerContainer = useRef<HTMLDivElement>(null);
+    const [trackScrolling, bodySize, containerRef] = useInfiniteControl(props);
+
     const observerCallback = useCallback(
         (entry: [IntersectionObserverEntry]) => {
             const [e] = entry;
@@ -31,12 +36,19 @@ export function StickyHeaderTable(props: PropsWithChildren): ReactElement {
             }
         };
     }, [headerContainer]);
+
     return (
-        <Fragment>
+        <div
+            className={classNames("sticky-table-container", className, { "infinite-loading": isInfinite })}
+            ref={containerRef}
+            onScroll={isInfinite ? trackScrolling : undefined}
+            role={props.role}
+            style={isInfinite && bodySize > 0 ? { ...props.style, maxHeight: bodySize } : props.style}
+        >
             <div className="sticky-sentinel" ref={headerContainer} style={{ height: 1, marginBottom: "-1px" }}></div>
             <div className="table" role="table">
-                {props.children}
+                {children}
             </div>
-        </Fragment>
+        </div>
     );
 }
